Added new method to create diagnostics
