+++ backend/src/fhir_service.py
@@ -17,7 +17,7 @@ from models.errors import (
     InconsistentIdError,
 )
 from models.fhir_immunization import ImmunizationValidator
-from models.utils.generic_utils import nhs_number_mod11_check, get_occurrence_datetime, get_disease_type
+from models.utils.generic_utils import nhs_number_mod11_check, get_occurrence_datetime, get_disease_type,create_diagnostics
 from models.utils.post_validation_utils import MandatoryError, NotApplicableError
 from pds_service import PdsService
 from s_flag_handler import handle_s_flag
@@ -102,7 +102,7 @@ class FhirService:
 
         patient = self._validate_patient(immunization)
         if "diagnostics" in patient: 
-                return (patient)
+                return (None,patient)
         try:
             imms = self.immunization_repo.update_immunization(imms_id, immunization, patient)
             return UpdateOutcome.UPDATE, Immunization.parse_obj(imms)
@@ -168,9 +168,10 @@ class FhirService:
         # TODO: is disease type a mandatory field? (I assumed it is)
         #  i.e. Should we provide a search option for getting Patient's entire imms history?
         if not nhs_number_mod11_check(nhs_number):
-                diagnostics_error = self.create_diagnostics(nhs_number)
+                diagnostics_error = create_diagnostics(nhs_number)
                 return (diagnostics_error)
         resources = self.immunization_repo.find_immunizations(nhs_number)
+        print(f"nhs_number :{nhs_number}")
         resources = [
             r for r in resources
             if FhirService.has_valid_disease_type(r, disease_types)
@@ -180,8 +181,9 @@ class FhirService:
         patient = self.pds_service.get_patient_details(nhs_number) if len(resources) > 0 else None
         if patient:
             pds_nhs_number = patient["identifier"][0]["value"]
+            print(f"PDS_nhs_number :{pds_nhs_number}")
             if pds_nhs_number != nhs_number:
-                diagnostics_error = self.create_diagnostics(nhs_number)
+                diagnostics_error = create_diagnostics(nhs_number)
                 return diagnostics_error
         entries = [
                 BundleEntry(
@@ -229,20 +231,9 @@ class FhirService:
             pds_nhs_number = patient["identifier"][0]["value"]
             print(f"PDS_NHS_NUMBER :{pds_nhs_number}")
             if pds_nhs_number != nhs_number :
-                    diagnostics=f"NHS Number: {nhs_number} is invalid or it doesn't exist."
-                    exp_error = {
-                                "diagnostics": diagnostics
-                                }
-                    return (exp_error)
+                    diagnostics_error = create_diagnostics(nhs_number)
+                    return diagnostics_error
         
             return patient
 
         raise InvalidPatientId(patient_identifier=nhs_number)
-    
-    
-    def create_diagnostics(nhs_number):
-                diagnostics=f"NHS Number: {nhs_number} is invalid or it doesn't exist."
-                exp_error = {
-                             "diagnostics": diagnostics
-                            }
-                return (exp_error)
